remove not used classes
