+++ src/ExadelMentorship.BusinessLogic/Features/WeatherFeature/CurrentWeather/CurrentWeatherCommand.cs
@@ -3,7 +3,7 @@ using ExadelMentorship.BusinessLogic.Interfaces;
 
 namespace ExadelMentorship.BusinessLogic.Features.WeatherFeature
 {
-    public class CurrentWeatherCommand : ICommand2 <CurrentWeatherCommandResponse>
+    public class CurrentWeatherCommand : ICommand <CurrentWeatherCommandResponse>
     {
         public string CityName { get; set; }
     }
