+++ src/ExadelMentorship.BusinessLogic/Features/WeatherFeature/MaxWeather/MaxWeatherCommand.cs
@@ -5,7 +5,7 @@ using System.Text;
 
 namespace ExadelMentorship.BusinessLogic.Features.WeatherFeature.MaxWeather
 {
-    public class MaxWeatherCommand : ICommand2<MaxWeatherCommandResponse>
+    public class MaxWeatherCommand : ICommand<MaxWeatherCommandResponse>
     {
         public string Cities { get; set; }
     }
