--- src/ExadelMentorship.BusinessLogic/Interfaces/ICommand2.cs
@@ -1,8 +1,0 @@
-﻿using System.Threading.Tasks;
-
-namespace ExadelMentorship.BusinessLogic.Interfaces
-{
-    public interface ICommand2<out TResult> 
-    {
-    }
-}
