+++ src/ExadelMentorship.BusinessLogic/MainJob.cs
@@ -1,10 +1,13 @@
 ﻿using ExadelMentorship.BusinessLogic.Exceptions;
 using ExadelMentorship.BusinessLogic.Features;
 using ExadelMentorship.BusinessLogic.Features.WeatherFeature;
+using ExadelMentorship.BusinessLogic.Features.WeatherFeature.CurrentWeather;
 using ExadelMentorship.BusinessLogic.Features.WeatherFeature.FutureWeather;
 using ExadelMentorship.BusinessLogic.Features.WeatherFeature.MaxWeather;
-using ExadelMentorship.BusinessLogic.Interfaces;
+using ExadelMentorship.BusinessLogic.Models;
 using System;
+using System.Collections.Generic;
+using System.Linq;
 using System.Threading.Tasks;
 
 namespace ExadelMentorship.BusinessLogic
@@ -33,7 +36,8 @@ namespace ExadelMentorship.BusinessLogic
                 {
                      try
                      {
-                        await _commandInvoker.Invoke(ParseCommand(command));
+                        var result = await _commandInvoker.Invoke(ParseCommand(command));
+                        PrintResult(result);
                         _rwOperation.ReadLine();
                      } 
                       catch (NotFoundException ex)
@@ -47,6 +51,79 @@ namespace ExadelMentorship.BusinessLogic
                 }
             } while (condition);
         }
+
+        private void PrintResult(dynamic result)
+        {
+            if (result is CurrentWeatherCommandResponse)
+            {
+                _rwOperation.WriteLine(Texts.CurrentWeatherCommandResponse,result.Name, result.Temperature, result.Comment);
+            }
+
+            if (result is IEnumerable<City>)
+            {
+                foreach (var city in result.cityList)
+                {
+                    _rwOperation.WriteLine($"Day {city.Date.ToString("dd/MM/yyyy")}: {city.Temperature}. {city.Comment}");
+                }
+              
+            }
+
+            if (result is MaxWeatherCommandResponse)
+            {
+                if (result.SuccessfulRequests > 0)
+                {
+                    _rwOperation.WriteLine(Texts.SuccessfulRequest, 
+                        result.MaxTempCityInfo.Temperature, 
+                        result.MaxTempCityInfo.Name,
+                        result.SuccessfulRequests, 
+                        result.FailedRequests, 
+                        result.CancelledRequests);
+
+                    if (result.Statistic)
+                    {
+                        DebugInfoProvider(result.MaxTempCityInfoList);
+                    }                    
+                }
+                else
+                {
+                    _rwOperation.WriteLine(Texts.NoSuccessful, result.FailedRequests, result.CancelledRequests);
+                }
+            }
+        }
+
+        private dynamic ParseCommand(int commnad)
+         {
+            if (commnad == 1)
+            {
+                _rwOperation.WriteLine("Please enter the city Name:");
+                return new CurrentWeatherCommand
+                {
+                    CityName = _rwOperation.ReadLine()
+                };
+            }
+            if (commnad == 2)
+            {
+                FutureWeatherCommand futureWeatherCommand = new FutureWeatherCommand();
+                
+                _rwOperation.WriteLine("Please enter the city Name:");
+                futureWeatherCommand.CityName = _rwOperation.ReadLine();
+
+                _rwOperation.WriteLine("Please enter interested days quantity:");
+                futureWeatherCommand.DayQuantity = _rwOperation.ReadLine();
+                
+                return futureWeatherCommand;
+            }
+            if (commnad == 3)
+            {
+                _rwOperation.WriteLine("Please enter the cities:");
+                return new MaxWeatherCommand
+                {
+                    Cities = _rwOperation.ReadLine(),
+                };
+            }
+            throw new NotImplementedException();
+        }
+
         private int GetActionFromUser()
         {
             string inputedLine;
@@ -61,21 +138,22 @@ namespace ExadelMentorship.BusinessLogic
             } while (!(inputedLine.Equals("0") || inputedLine.Equals("1") || inputedLine.Equals("2") || inputedLine.Equals("3")));
             return Convert.ToInt32(inputedLine);
         }
-        private ICommand ParseCommand(int commnad)
-         {
-            //if (commnad == 1)
-            //{
-            //    return new CurrentWeatherCommand();
-            //}
-            //if (commnad == 2)
-            //{
-            //    return new FutureWeatherCommand();
-            //}
-            //if (commnad == 3)
-            //{
-            //    return new MaxWeatherCommand();
-            //}
-            throw new NotImplementedException();
+
+        private void DebugInfoProvider(IEnumerable<MaxTempCityInfo> maxTempCityInfoList)
+        {
+            var lits = maxTempCityInfoList.ToList();
+            foreach (var item in maxTempCityInfoList)
+            {
+                if (item.Name != null)
+                {
+                    _rwOperation.WriteLine(Texts.DebugInfo, item.Name, item.Temperature, item.DurationTime);
+                }
+                if (item.ErrorMessage != null)
+                {
+                    _rwOperation.WriteLine(item.ErrorMessage);
+                }
+            }
+
         }
 
     }
