+++ src/ExadelMentorship.ConsoleApp/DI.cs
@@ -22,12 +22,12 @@ namespace ExadelMentorship.IntegrationTests
 
             return new ServiceCollection()
                 .AddSingleton(config)
-                .AddSingleton<MainJob2>()
-                .AddSingleton<CommandInvoker2>()
+                .AddSingleton<MainJob>()
+                .AddSingleton<CommandInvoker>()
                 .AddSingleton<IWeatherApiService, WeatherApiService>()
-                .AddSingleton<ICommandHandler2<CurrentWeatherCommand, CurrentWeatherCommandResponse>, CurrentWeatherCommandHandler>()               
-                .AddSingleton<ICommandHandler2<FutureWeatherCommand, IEnumerable<City>>, FutureWeatherCommandHandler>()
-                .AddSingleton<ICommandHandler2<MaxWeatherCommand, MaxWeatherCommandResponse>, MaxWeatherCommandHandler>()
+                .AddSingleton<ICommandHandler<CurrentWeatherCommand, CurrentWeatherCommandResponse>, CurrentWeatherCommandHandler>()               
+                .AddSingleton<ICommandHandler<FutureWeatherCommand, IEnumerable<City>>, FutureWeatherCommandHandler>()
+                .AddSingleton<ICommandHandler<MaxWeatherCommand, MaxWeatherCommandResponse>, MaxWeatherCommandHandler>()
                 .AddSingleton<IRWOperation, ConsoleOperation>()
                 .AddHttpClient()
                 .BuildServiceProvider()
